pricing.js added
In this pricing.js file is helps to pricing.html page, specifically a pricing section. It first retrieves elements from the html, such as the normal and hidden containers, and a toggle checkbox. The script then checks the local storage for a boolean value indicating whether the checkbox was previously checked, defaulting to false if no value is found. The checkbox state and visibility of the containers are then set based on this value. When the checkbox state changes, the script updates the local storage, toggles the visibility of the normal and hidden containers, allowing users to switch between annual and monthly pricing options.